Permit customizing ignored exceptions when polling job status (#98)
The actual code change here is very small -- the bulk of the changes
is new/edited test code.

Before this patch, starting a blocking job would ignore any HTTP
errors when checking for job status to complete, until polling finally
timed out. This change allows a user to specify which exceptions to
ignore by passing a tuple (which is the data type required by the
polling2 library) of ignored exceptions. Providing an empty tuple
means that no exceptions are ignored.

The original default behavior was not changed.

On the test side:

- a new unit test file was added: test_jobs.py
- fixtures were moved to a new conftest.py file to be more
  accessible across different tests

Co-authored-by: David Knupp <david.knupp@dominodatalab.com>